+++ packages/imperative/src/rest/__tests__/client/AbstractRestClient.unit.test.ts
@@ -13,6 +13,7 @@ import * as fs from "fs";
 import * as https from "https";
 import * as http from "http";
 import { Session } from "../../src/session/Session";
+import { AuthOrder } from "../../src/session/AuthOrder";
 import {
     AUTH_TYPE_BASIC, AUTH_TYPE_BEARER, AUTH_TYPE_CERT_PEM, AUTH_TYPE_NONE, AUTH_TYPE_TOKEN
 } from "../../src/session/SessConstants";
@@ -43,6 +44,9 @@ import { EnvironmentalVariableSettings } from "../../../imperative";
 
 describe("AbstractRestClient tests", () => {
     let setPasswordAuthSpy: any;
+
+    // never run putTopAuthInSession. It has its own unit test.
+    AuthOrder.putTopAuthInSession = jest.fn();
 
     beforeEach(() => {
         jest.restoreAllMocks();